requirements.txt
backcall==0.1.0
decorator==4.4.0
entrypoints==0.3
flake8==3.7.6
ipython==7.3.0
ipython-genutils==0.2.0
jedi==0.13.3
mccabe==0.6.1
mysql-connector==2.1.6
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.0
Pygments==2.3.1
PyYAML==5.1
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
